Sépare standardisateur et standardisateur glissant
