+++ app/models/restitution/standardisateur.rb
@@ -2,31 +2,6 @@
 
 module Restitution
   class Standardisateur
-    def initialize(metriques, collect_metriques)
-      @metriques = metriques
-      @collect_metriques = collect_metriques
-    end
-
-    def moyenne_metrique(metrique)
-      aggrege_metrique(:average, metrique)
-    end
-
-    def ecart_type_metrique(metrique)
-      aggrege_metrique(:stddev_pop, metrique)
-    end
-
-    def moyenne_metriques
-      @moyenne_metriques ||= @metriques.each_with_object({}) do |metrique, memo|
-        memo[metrique] = moyenne_metrique(metrique)
-      end
-    end
-
-    def ecart_type_metriques
-      @ecart_type_metriques ||= @metriques.each_with_object({}) do |metrique, memo|
-        memo[metrique] = ecart_type_metrique(metrique)
-      end
-    end
-
     def standardise(metrique, valeur)
       return if valeur.nil? || ecart_type_metriques[metrique].nil?
 
@@ -37,16 +12,6 @@ module Restitution
           (valeur - moyenne_metriques[metrique]) / ecart_type_metriques[metrique]
         )
       end
-    end
-
-    private
-
-    def aggrege_metrique(fonction, metrique)
-      @collect_metriques
-        .call
-        .where.not(metriques: {})
-        .calculate(fonction, "(metriques ->> '#{metrique}')::numeric")
-        .to_f
     end
   end
 end
